- web, tablet, mobile별 레이아웃 - window size를 detect 하는 `vueuse` 활용 - 추후 email 작성 폼 작성 예정
