+++ src/assets/variables.scss
@@ -1,3 +1,13 @@
+body {
+    margin: 0px;
+    padding: 0px;
+}
+
+body * {
+    box-sizing: border-box;
+    border: 0px solid #e5e7eb;
+}
+
 body * {
     --button-base-border-radius: 12px;
 }
